16대 이하 법안 크롤링 예외처리

+++ bills/specific/html.py
@@ -4,34 +4,48 @@
 import os
 import sys
 import urllib2
-import pandas as pd
+from operator import itemgetter
 
 from settings import BASEURL, DIR, HTML_FIELDS
 import utils
 
 def get_metadata(assembly_id, range=(None, None)):
     with open('%s/%d.csv' % (DIR['meta'], assembly_id), 'r') as f:
-        data = pd.read_csv(f)[range[0]:range[1]]
+        data = []
+        for row in f.readlines():
+            items = [item.strip('"') for item in row.decode('utf-8').split(',')]
+            data.append(itemgetter(0,3,8)(items))
 
     meta = {}
-    for d in zip(data['bill_id'], data['link_id'], data['has_summaries']):
+    for d in data:
         meta[d[0]] = (d[1], d[2])
 
     return meta
 
 def get_page(assembly_id, bill_id, link_id, field):
+    url = '%s%s' % (BASEURL[field], link_id)
     outp = '%s/%s/%s.html' % (DIR[field], assembly_id, bill_id)
 
-    is_first = True
-    url = '%s%s' % (BASEURL[field], link_id)
+    if not os.path.isfile(outp):
+        is_first = True
+        while is_first or 'TEXTAREA ID="MSG" STYLE="display:none"' in doc:
+            r = urllib2.urlopen(url)
+            doc = r.read()
+            is_first = False
 
-    while is_first or 'TEXTAREA ID="MSG" STYLE="display:none"' in doc:
-        r = urllib2.urlopen(url)
-        doc = r.read()
-        is_first = False
+        with open(outp, 'w') as f:
+            f.write(doc)
 
-    with open(outp, 'w') as f:
-        f.write(doc)
+def get_specifics(assembly_id, bill_id, link_id):
+    if assembly_id > 16:
+        baseurl = BASEURL['specifics']
+    else:
+        baseurl = BASEURL['specifics_old']
+
+    outp = '%s/%s/%s.html' % (DIR['specifics'], assembly_id, bill_id)
+    if not os.path.isfile(outp):
+        url = '%s%s' % (baseurl, link_id)
+        utils.get_webpage(url, outp)
 
 def get_summaries(assembly_id, bill_id, link_id, has_summaries):
     if has_summaries==1:
@@ -47,8 +61,9 @@ def get_html(assembly_id, range=(None, None)):
 
     for bill_id in metadata:
         link_id, has_summaries = metadata[bill_id]
-        for field in HTML_FIELDS[:-1]:
+        for field in HTML_FIELDS[1:3]:
             get_page(assembly_id, bill_id, link_id, field)
+        get_specifics(assembly_id, bill_id, link_id)
         get_summaries(assembly_id, bill_id, link_id, has_summaries)
 
         sys.stdout.write('%s\t' % bill_id)
